# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List all subjects
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a custom subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Created subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /subjects/{id}:
    get:
      operationId: getSubject
      tags: [subjects]
      summary: Get a subject by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /quizzes:
    get:
      operationId: listQuizzes
      tags: [quizzes]
      summary: List quizzes
      parameters:
        - name: mode
          in: query
          required: false
          schema:
            type: string
            enum: [placement, practice, exam]
        - name: subjectId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizSummary"
  /quizzes/generate:
    post:
      operationId: generateQuiz
      tags: [quizzes]
      summary: Generate a quiz, test, or exam with AI
      description: |
        Generates a quiz from a chosen subject (placement), an uploaded document,
        or a free-form topic. Returns the full quiz including questions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizGenerateInput"
      responses:
        "201":
          description: Generated quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /quizzes/{id}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      summary: Get a quiz with its questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteQuiz
      tags: [quizzes]
      summary: Delete a quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Subject:
      type: object
      required: [id, name, slug, category, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        icon:
          type: ["string", "null"]
        isCustom:
          type: boolean
        createdAt:
          type: string
          format: date-time
    SubjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        category:
          type: string
...
    Question:
      type: object
      required: [id, prompt, options, correctIndex, order]
      properties:
        id:
          type: integer
        prompt:
          type: string
        options:
          type: array
          items:
            type: string
        correctIndex:
          type: integer
        explanation:
          type: ["string", "null"]
        order:
          type: integer
    Quiz:
      type: object
      required: [id, title, mode, difficulty, questionCount, createdAt, questions]
      properties:
        id:
          type: integer
        title:
          type: string
        mode:
          type: string
          enum: [placement, practice, exam]
        subjectId:
          type: ["integer", "null"]
        subjectName:
          type: ["string", "null"]
        documentId:
          type: ["integer", "null"]
        topic:
          type: ["string", "null"]
        difficulty:
          type: string
        questionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
    QuizSummary:
      type: object
      required: [id, title, mode, difficulty, questionCount, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        mode:
          type: string
          enum: [placement, practice, exam]
        subjectId:
          type: ["integer", "null"]
        subjectName:
          type: ["string", "null"]
        topic:
          type: ["string", "null"]
        difficulty:
          type: string
        questionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    QuizGenerateInput:
      type: object
      required: [mode]
      properties:
        mode:
          type: string
          enum: [placement, practice, exam]
        subjectId:
          type: integer
        documentId:
          type: integer
        topic:
          type: string
        title:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, mixed]
        questionCount:
          type: integer
          minimum: 3
          maximum: 25
...