# lib/api-zod/src/generated/types/quiz.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Question } from './question';
import type { QuizMode } from './quizMode';

export interface Quiz {
  id: number;
  title: string;
  mode: QuizMode;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  /** @nullable */
  documentId?: number | null;
  /** @nullable */
  topic?: string | null;
  difficulty: string;
  questionCount: number;
  createdAt: Date;
  questions: Question[];
}