# lib/api-zod/src/generated/types/quizSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuizSummaryMode } from './quizSummaryMode';

export interface QuizSummary {
  id: number;
  title: string;
  mode: QuizSummaryMode;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  /** @nullable */
  topic?: string | null;
  difficulty: string;
  questionCount: number;
  createdAt: Date;
}